% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit1.R
\name{stsplit1}
\alias{stsplit1}
\title{split a string}
\usage{
stsplit1(x, split)
}
\arguments{
\item{x}{A character vector with one element}

\item{split}{What to split on.}
}
\value{
A charcter vector.
}
\description{
split a string
}
